made viewport scroll with player movement

+++ game/position.go
@@ -1,5 +1,5 @@
 package game
 
 type Position struct {
-	Xpos, Ypos float64
+	Xpos, Ypos int
 }
